+++ src/configs/node-dependencies.ts
@@ -9,6 +9,9 @@ export interface NodeDependenciesEslintConfigOptions extends UnConfigOptions<'no
    * - `true`: enforces to use the absolute version only on `dependencies` and `devDependencies`.
    * - `'never'`: enforces not to use the absolute version.
    * - `false`: do not enforce anything.
+   *
+   * Affected rules:
+   * - [`absolute-version`](https://ota-meshi.github.io/eslint-plugin-node-dependencies/rules/absolute-version.html)
    * @default false
    */
   enforceAbsoluteVersion?: